requirements.txt
slack_bolt==1.23.0
slack_cli_hooks==0.1.0
slack_sdk==3.36.0
anthropic==0.60.0
python-dotenv==1.1.1
requests==2.32.4
claude-code-sdk==0.0.19

pytest==8.4.1
flake8==7.3.0
black==25.1.0
